+++ src/components/common/PreviewCard.js
@@ -2,7 +2,7 @@ import React from "react"
 import PropTypes from "prop-types"
 import styled from "styled-components"
 
-import FlexPad from "../basics/FlexPad"
+import FlexPad from "../basic/FlexPad"
 import { colors } from "../../constants/common"
 
 const Wrapper = styled.div`
